requirements.txt
torch
transformers
datasets
huggingface_hub
requests
python-dotenv
